--- examples/udp_tag.cpp
@@ -1,140 +1,0 @@
-/*
-For ESP32 UWB or ESP32 UWB Pro
-*/
-
-#include <SPI.h>
-#include <DW1000Ranging.h>
-#include <WiFi.h>
-#include "link.h"
-
-#ifdef MAKERFABS
-
-#define SPI_SCK 18
-#define SPI_MISO 19
-#define SPI_MOSI 23
-#define DW_CS 4
-#define PIN_RST 27
-#define PIN_IRQ 34
-
-#endif
-
-#ifdef BLUEPILL
-
-#define SPI_SCK PB13
-#define SPI_MISO PB14
-#define SPI_MOSI PB15
-
-// connection pins
-const uint8_t PIN_RST = PA12; // reset pin
-const uint8_t PIN_IRQ = PA11; // irq pin
-const uint8_t PIN_SS = PB12;   // spi select pin
-
-#endif
-
-const char *ssid = "xxxx";
-const char *password = "xxxxx";
-const char *host = "xxxx";
-WiFiClient client;
-
-struct MyLink *uwb_data;
-int index_num = 0;
-long runtime = 0;
-String all_json = "";
-
-void newRange()
-{
-    char c[30];
-
-    Serial.print("from: ");
-    Serial.print(DW1000Ranging.getDistantDevice()->getShortAddress(), HEX);
-    Serial.print("\t Range: ");
-    Serial.print(DW1000Ranging.getDistantDevice()->getRange());
-    Serial.print(" m");
-    Serial.print("\t RX power: ");
-    Serial.print(DW1000Ranging.getDistantDevice()->getRXPower());
-    Serial.println(" dBm");
-    fresh_link(uwb_data, DW1000Ranging.getDistantDevice()->getShortAddress(), DW1000Ranging.getDistantDevice()->getRange(), DW1000Ranging.getDistantDevice()->getRXPower());
-}
-
-void newDevice(DW1000Device *device)
-{
-    Serial.print("ranging init; 1 device added ! -> ");
-    Serial.print(" short:");
-    Serial.println(device->getShortAddress(), HEX);
-
-    add_link(uwb_data, device->getShortAddress());
-}
-
-void inactiveDevice(DW1000Device *device)
-{
-    Serial.print("delete inactive device: ");
-    Serial.println(device->getShortAddress(), HEX);
-
-    delete_link(uwb_data, device->getShortAddress());
-}
-
-void send_udp(String *msg_json)
-{
-    if (client.connected())
-    {
-        client.print(*msg_json);
-        Serial.println("UDP send");
-    }
-}
-
-void setup()
-{
-    Serial.begin(115200);
-
-    WiFi.mode(WIFI_STA);
-    WiFi.setSleep(false);
-    WiFi.begin(ssid, password);
-    while (WiFi.status() != WL_CONNECTED)
-    {
-        delay(500);
-        Serial.print(".");
-    }
-    Serial.println("Connected");
-    Serial.print("IP Address:");
-    Serial.println(WiFi.localIP());
-
-    if (client.connect(host, 80))
-    {
-        Serial.println("Success");
-        client.print(String("GET /") + " HTTP/1.1\r\n" +
-                     "Host: " + host + "\r\n" +
-                     "Connection: close\r\n" +
-                     "\r\n");
-    }
-
-    delay(1000);
-
-    //init the configuration
-    #ifdef MAKERFABS
-    SPI.begin(SPI_SCK, SPI_MISO, SPI_MOSI);
-    #endif
-
-    #ifdef BLUEPILL
-    SPI.begin(115200);
-    #endif
-    DW1000Ranging.initCommunication(PIN_RST, DW_CS, PIN_IRQ);
-    DW1000Ranging.attachNewRange(newRange);
-    DW1000Ranging.attachNewDevice(newDevice);
-    DW1000Ranging.attachInactiveDevice(inactiveDevice);
-
-    //we start the module as a tag
-    DW1000Ranging.startAsTag("7D:00:22:EA:82:60:3B:9C", DW1000.MODE_LONGDATA_RANGE_LOWPOWER);
-
-    uwb_data = init_link();
-}
-
-void loop()
-{
-    DW1000Ranging.loop();
-    if ((millis() - runtime) > 1000)
-    {
-        make_link_json(uwb_data, &all_json);
-        send_udp(&all_json);
-        runtime = millis();
-    }
-}
