+++ platformio.ini
@@ -44,6 +44,27 @@ build_flags = '-w'
               '-D MAKERFABS'
               '-D UWB_ANCHOR'
 
+[env:esp_anchor_ad]
+# extends doesn't work, great job platformio. That means
+# we have to duplicate all the environment lines.
+# TODO(michalc): figure out why extends doesn't work
+extends = esp_tag
+platform = espressif32
+board = esp-wrover-kit
+framework = arduino
+#lib_deps = thotro/DW1000@^0.9
+upload_speed = 921600
+board_build.f_flash = 80000000L
+board_build.flash_mode = qio
+board_build.partitions = default.csv
+upload_port = /dev/ttyUSB0
+monitor_port = /dev/ttyUSB0
+monitor_speed = 115200
+build_flags = '-w'
+              '-D MAKERFABS'
+              '-D UWB_ANCHOR'
+			  '-D UWB_ANCHOR_AD'
+
 
 [env:bp_tag]
 platform = ststm32
@@ -74,3 +95,21 @@ monitor_port = /dev/ttyUSB0
 build_flags = '-w'
               '-D BLUEPILL'
               '-D UWB_ANCHOR'
+
+[env:bp_anchor_ad]
+# extends doesn't work, great job platformio. That means
+# we have to duplicate all the environment lines.
+# TODO(michalc): figure out why extends doesn't work
+extends = bp_tag
+platform = ststm32
+board = bluepill_f103c8
+framework = arduino
+#lib_deps = thotro/DW1000@^0.9
+#upload_flags = -c set CPUTAPID 0x2ba01477
+upload_protocol = stlink
+monitor_speed = 115200
+monitor_port = /dev/ttyUSB0
+build_flags = '-w'
+              '-D BLUEPILL'
+              '-D UWB_ANCHOR'
+			  '-D UWB_ANCHOR_AD'
